refactor(live): extract exit pipeline into LiveExitCoordinator (#486) (#815)
Mirror of the entry extraction (#813). Move _check_exit_conditions,
_execute_exit, and _execute_exit_locked verbatim into a new
LiveExitCoordinator (engines/live/execution/exit_coordinator.py) via a
mechanical self. -> state. rewrite against a LiveExitEngineState Protocol
backref. The engine keeps thin delegating wrappers so all call sites and
test mock points are unchanged.

Routing: check_exit_conditions invokes the close through the engine's
_execute_exit wrapper (state._execute_exit) so engine-level test mocks
(engine._execute_exit = Mock()) still intercept; execute_exit ->
execute_exit_locked is a coordinator-internal call. The base-asset close
lock (#703, re-entrant for an entry's failed-SL emergency close) and the
resting-stop cancel-before-close ordering (#710) are preserved.

Pure refactor: backtest determinism fingerprint byte-identical before/after;
trading_engine.py 3,574 -> 3,130 lines. trade_close_accounting helpers kept
as a re-export from trading_engine (tests import them from there).
