+++ docs/live_trading.md
@@ -1,6 +1,6 @@
 # Live trading
 
-> **Last Updated**: 2026-06-10
+> **Last Updated**: 2026-06-15
 > **Related Documentation**: [Backtesting](backtesting.md), [Monitoring](monitoring.md), [Database](database.md)
 
 `src/engines/live/trading_engine.py` powers the real-time execution stack. It shares core building blocks with the backtester while adding
@@ -34,6 +34,7 @@ continuous polling, account synchronisation, and resilience features required fo
 | `StrategyHotSwapCoordinator` | `engines/live/strategy_hot_swap.py` | Public `hot_swap_strategy`/`update_model`, StrategyManager callbacks, loop-applied pending-update application, post-swap refresh of trailing-stop/partial-ops/time-exit policies + component risk re-binding | None — entry points/callbacks only queue a `StrategyManager`-locked pending update (caller thread); all engine-state mutation runs in `apply_pending_strategy_update` on the trading-loop thread |
 | `WebSocketHealthMonitor` | `engines/live/ws_health.py` | WS stream startup, the background health-monitor thread, kline/user-stream staleness + reconnect/probe decisions, degraded-user hard-reconnect, draining the order-fill exit queue on the loop thread | None — the monitor holds no locks. It owns a single daemon thread (handle on the engine, `state._ws_health_thread`) and writes its reconnect-failure counters / `_ws_kline_active` flag only from that one thread (GIL-atomic single-writer); the trading loop reads them. Cross-thread order-fill handoff goes through the thread-safe `_pending_fill_exits` `SimpleQueue`, drained on the loop thread. Provider-owned WS connection state is read-only here |
 | `LiveEntryCoordinator` | `engines/live/execution/entry_coordinator.py` | Entry decision (signal/sizing/SL-TP derivation) + base-asset-locked order execution: duplicate/limit guards, balance+fee accounting, position tracking, risk re-registration, stop-loss placement, emergency-close fallbacks | Serialises each entry on the symbol's base-asset lock (`state._base_asset_locks`) across submit → track → SL placement; the lock lives on the engine and the emergency-close fallback re-acquires it re-entrantly via `state._execute_exit` (#703). Runs on the trading-loop thread; all engine state mutated through the backref as before |
+| `LiveExitCoordinator` | `engines/live/execution/exit_coordinator.py` | Exit decision (`check_exit_conditions`: per-position SL/TP/runtime evaluation + strategy-execution logging) and base-asset-locked close (`execute_exit` → `execute_exit_locked`): resting-stop cancel-before-close (#710), realized-PnL + margin-interest accounting, balance update, trade persistence/CLOSED flip (#657), re-protect on failed close | Serialises each close on the symbol's base-asset lock (`state._base_asset_locks`, #703); the lock lives on the engine and is re-entrant (an entry's failed-SL emergency close routes through the engine's `_execute_exit` wrapper on the same thread). `check_exit_conditions` invokes the close via `state._execute_exit` (the engine wrapper) so engine-level test mocks still intercept. Runs on the trading-loop thread; all engine state mutated through the backref as before |
 | `LivePositionTracker` | `engines/live/execution/position_tracker.py` | Position state | Owns `_positions_lock`; `positions` property returns a defensive copy |
 | `OrderTracker` | `engines/live/order_tracker.py` | Order fill polling + callbacks | Owns its internal lock; engine callbacks defer closes to the trading loop via `_pending_fill_exits` |
 | `SharedEntryHandlerMixin` | `engines/shared/execution/entry_handler_mixin.py` | Entry-plan extraction + dynamic-risk sizing, identical for backtest and live | None — delegates to shared `DynamicRiskHandler` |
